time_operations: add cc
